+++ packages/framework-provider-local/src/library/read-model-adapter.ts
@@ -7,6 +7,7 @@ import {
   UUID,
 } from '@boostercloud/framework-types'
 import { ReadModelRegistry } from '../services/read-model-registry'
+import { QueryValue, filterToQuery } from './searcher-adapter'
 
 export async function rawReadModelEventsToEnvelopes(
   config: BoosterConfig,
@@ -67,53 +68,4 @@ export async function searchReadModel(
   const result = await db.query(query)
   logger.debug('[ReadModelAdapter#searchReadModel] Search result: ', result)
   return result?.map((envelope) => envelope.value) ?? []
-}
-
-type QueryValue = number | string | boolean
-type QueryOperation<TValue> =
-  | TValue
-  | {
-      [TKey in '$lt' | '$lte' | '$gt' | '$gte' | '$ne']?: TValue
-    }
-  | {
-      [TKey in '$in' | '$nin']?: Array<TValue>
-    }
-  | {
-      [TKey in '$regex' | '$nin']?: RegExp
-    }
-
-const queryOperatorTable: Record<string, (vals: Array<QueryValue>) => QueryOperation<QueryValue>> = {
-  '=': (val) => val[0],
-  '!=': (val) => ({ $ne: val[0] }),
-  '<': (val) => ({ $lt: val[0] }),
-  '>': (val) => ({ $gt: val[0] }),
-  '<=': (val) => ({ $lte: val[0] }),
-  '>=': (val) => ({ $gte: val[0] }),
-  in: (val) => ({ $in: val }),
-  between: (val) => ({ $gt: val[0], $lte: val[1] }),
-  contains: buildRegexQuery.bind(null, 'contains'),
-  'not-contains': buildRegexQuery.bind(null, 'not-contains'),
-  'begins-with': buildRegexQuery.bind(null, 'begins-with'),
-}
-
-function buildRegexQuery(operation: string, vals: Array<QueryValue>): QueryOperation<QueryValue> {
-  let matcher = vals[0]
-  if (typeof matcher != 'string') {
-    throw new Error(`Attempted to perform a ${operation} operation on a non-string`)
-  }
-  if (operation === 'not-contains') {
-    throw new Error('not-contains not implemented yet')
-  }
-  if (operation === 'begins-with') {
-    matcher = `^${matcher}`
-  }
-  return { $regex: new RegExp(matcher) }
-}
-
-/**
- * Transforms a GraphQL Booster filter into an neDB query
- */
-function filterToQuery(filter: Filter<QueryValue>): QueryOperation<QueryValue> {
-  const query = queryOperatorTable[filter.operation]
-  return query(filter.values)
 }
